Markdown formatting and minor grammar fixes
This is a bit noisy but does not represent very many substantive fixes. The changes here are of two types:
* Small grammar fixes.
* Introducing linebreaks for every paragraph to make it easier to read diffs and make patches for content.

The former are small (eg => e.g.; which => that; linking out to references; see => refer to as inclusive of people with sight impairment), so I'm including them here rather than doing a two-step of refactor markdown, then make substantive changes.

The latter refactoring is a common convention with Markdown, AsciiDoc, et al to make it easier to read the diff for a languge/copyedit. Unlike code, changes to content are not as easy to spot, especially when in the middle of a paragraph of three to six sentences.

The refactoring here is to put each sentence on its own line, then use the empty line to signify a new paragraph. Numbering works similar, just with the number instead of the empty line representing the new paragraph.