Refactoring Markdown to consistent format
As per #4, the idea here is to make it easier to copyedit and read the diffs in a collaborative document by putting each sentence on a line of its own. The paragraph remains in the rendering, but the source is easier to read and maintain.

Snuck in a few minor edits as well -- a punctuation consistency fix, capitalization fix, and making it clear that Langston Hughes did not author any blog posts. :)